Removes webhook logic from gitlab handler class, moves
