+++ integrations/gitlab_v1/client.py
@@ -140,10 +140,7 @@ class GitlabHandler:
                 *[self._fetch_additional_data(resource, item) for item in response]
             )
 
-
-            for item in enhanced_items:
-                yield item
-
+            yield enhanced_items
 
             if len(response) < PAGE_SIZE:
                 logger.debug(f"Last page reached for resource '{resource}', no more data.")
@@ -158,61 +155,3 @@ class GitlabHandler:
     ) -> Dict[str, Any]:
         """Get a single resource by kind and ID."""
         return await self._send_api_request(f"{resource_kind}/{resource_id}")
-
-
-    async def create_webhook(self, project_id: str, webhook_url: str, webhook_token: str, events: List[str]) -> Dict[str, Any]:
-        """Create a webhook for a specific project."""
-        endpoint = f"projects/{project_id}/hooks"
-        data = {
-            "url": webhook_url,
-            "token": webhook_token,
-            "push_events": "push" in events,
-            "issues_events": "issues" in events,
-            "merge_requests_events": "merge_requests" in events,
-        }
-        return await self._send_api_request(endpoint, method="POST", json_data=data)
-
-
-    async def list_webhooks(self, project_id: str) -> List[Dict[str, Any]]:
-        """List all webhooks for a specific project."""
-        endpoint = f"projects/{project_id}/hooks"
-        return await self._send_api_request(endpoint)
-
-
-    async def update_webhook(self, project_id: str, webhook_id: int, webhook_url: str, webhook_token: str, events: List[str]) -> Dict[str, Any]:
-        """Update an existing webhook for a specific project."""
-        endpoint = f"projects/{project_id}/hooks/{webhook_id}"
-        data = {
-            "url": webhook_url,
-            "token": webhook_token,
-            "push_events": "push" in events,
-            "issues_events": "issues" in events,
-            "merge_requests_events": "merge_requests" in events,
-        }
-        return await self._send_api_request(endpoint, method="PUT", json_data=data)
-
-    async def setup_webhooks(self, webhook_url: str, webhook_token: str, events: List[str]) -> None:
-        """Set up webhooks for all accessible projects."""
-        async for page in self.get_paginated_resources("projects", params={"owned": True}):
-            for project in page:
-                if not isinstance(project, dict) or "id" not in project:
-                    logger.error(f"Invalid project structure: {project}")
-                    continue
-
-                project_id = str(project["id"])
-                logger.info(f"Processing project: {project_id}")
-
-                try:
-                    existing_webhooks = await self.list_webhooks(project_id)
-                    webhook_exists = any(
-                        isinstance(hook, dict) and hook.get("url") == webhook_url
-                        for hook in existing_webhooks
-                    )
-
-                    if not webhook_exists:
-                        await self.create_webhook(project_id, webhook_url, webhook_token, events)
-                        logger.info(f"Created webhook for project {project_id}")
-                    else:
-                        logger.info(f"Webhook already exists for project {project_id}")
-                except Exception as e:
-                    logger.error(f"Failed to set up webhook for project {project_id}: {str(e)}")
